add interupt lines (for NMI)
I wasn't planning on reworking interrupts today, but one thing led
to another, and... yeah.

All I really wanted to do today was get a Palette Table debug view,
but for some reason, it isn't being populated! I had a suspicion
that maybe implementing interrupts would make something click, but
alas, interrupts are in, and no data is being written to the
palette tables :(

I'm just going to leave it be for now, and hope that as more of the
PPU comes together, it will just start working.