work on sprite rendering
this isn't going well.

i've decided to forget accuracy for the moment, and to focus on
just getting the damn thing working with simple stuff, like the
Zelda title-screen demo, or Donkey Kong.

But fuck me man, this shit is complicated.

I'm going to bed now, and hopefully, next time I work on this,
i'll be able to get sprite rendering working properly.

hell, maybe i'll just scrap that, and instead focus on rendering
the background properly. I have a feeling that's going to be
a bit easier...

Right now, I really want to give up, and call it quits on this
project, because holy crap am I getting annoyed.
But then I remember that one day soon, everything is going to
work, and all this effort will have been worth it.

I hope that day comes soon.