trying random stuff to get things working
let's start with the good:

- well, graphics are a *bit* less janky now, since I fixed the bgr
fetch timings a bit. so that's nice.
- I found out that I forgot to emulate a instruction (CLI). (darn
NESTEST didn't ever test that one!)

and, well, that's it. shit's still fucked yo.

as a hail mary, i'm going to look into emulating some basic APU
stuff, hoping that it will magically make the timings align.

also, i'm starting to get a sneaking suspicion that my fancy
operator overloading and template magic shenanigans are breaking
things in very subtle ways. that fucking terrifies me. it's going
to be a massive shitshow to debug if that is the case.

anyhoo, that's where i'm at right now. time to get some dinner :)