fixed MMC1 enough to boot Zelda!
You know what the issue was?
The ram enable flag in the MMC1 control register enables access to
SRAM when it is set to 0.
Let me repeat that: 0 == enabled. 1 == disabled.

WHY WOULD ANY SANE PERSON DO THAT GOD DAMMIT WHY AAAAAGGHHHHH!

...

Sorry about that, it's just shit like that which really pisses me
off.

Well, regarless, now that i've put that fix in, Zelda actually
boots. Albeit with garbled graphics.

And other MMC1 games also still have the garbled-graphics issue,
which I still can't seem to fix...

As usual, i'm scared to say it but the issues probably lie within
the PPU. That said, I think i'm done with hammering on MMC1 for now
and instead I think i'll shift gears and see if I can get MMC3 up
and running. If graphical issue still remain with MMC3, then odds
are the issues are with the PPU, and not with MMC1.

weeeeeeeeeee I looooove emulator developmenttttt