cleanup serialization + fix savestates (save more data)
Save-states now also save the APU (which wasn't actually that hard
to do lol), and a better overview of Mapper state. I pushed the
`update_banks()` function up to the base mapper implementation,
and made it overridable, as by doing so I can override the
deserialize function on the mapper-level to call `update_banks`
after deserialization.

After hammering out the initial implementation, I took a look-back
and realized there was a lot that I could have done better. So,
with a fresh-mind, I cleaned up how serialization works under the
hood a bunch. It's a lot cleaner now, and AFAIK, it doesn't rely
on [that much] undefined behavior.

There are a few more things to iron out with it before i'm totally
happy with how it works, but i'm pretty proud of it right now :)