tried to implement Mapper 9 (Punch Out!!) + small .sav refactor
MMC2 is fancy since it snoops reads from the PPU to do bank-
switching on-the-fly, with 0 CPU intervention.
Unfortuantely, this means that for it to work properly, I need to
make my PPU fetching code accurate.

So yeah, if I ever fix the PPU, it should make Punch Out!! work :)