cleanups and refactors
One major thing I changed is that `nes.cc` no longer calls `new` to
instantiate components, and instead simply contains them. This is
nice for 2 reasons: 1) it is more faithful the the hardware
metaphor i'm going for, and 2) it lets me get a good idea of ANESE
Core's memory footprint, which is neat.

I also finally cleaned up the PPU's nametable mirroring, and while
it looks a bit uglier, at actual operations should be more readable
for those who know what to look for.