finished initial clenaup of SDL2 frontend
It wasn't that hard to do per-se, it's more about me finally
getting around to doing it. main.cc is hella clean, just making a
new SDL_GUI instance, and letting it RAII all of the SDL stuff.

Also, now that the new project directory structure is in place, it
shouldn't be that hard to add a LibRetro core! Fun!