attempted to add zapper support
it's not working though, and i'm not entirely sure why...
I doubt i'll spend more time looking into this right now though,
since it's a relatively minor thing. I was hoping it would
Just Work (TM), but alas, it looks to be a bit more involved...

Welp, it was worth a shot.

(oh, and if you can't tell, i'm just procrastinating starting to
reimplement the APU)