added hi and lo pass filters to apu
as someone who hasn't taken a signals processing course, figuring
out how to get these things working was a real pain.

Luckilly, Wikipedia is aweomse, and straight-up provides an
algorithm for both lopass and hipass filters. All it needed was a
bit of modification, and bam, easy as that!

----

As a sidenote, as i've been developing the APU, i've compared it to
some of the other APUs out there (including Blarrg's), and it
definately sounds shittier.
Maybe it's because i don't have DMC implemented yet? Unlikely...

Odds are i'll have to look into actual waveform synthesis if I want
it to sound better, which will be fun-on-the-bun.