improved parameter passing into ANESE core
no more janky DEBUG_VARS! Now, there is a defined NES_Params object
that is passed (by const ref) to the NES on construction, which is
passed along to all components that need it, with each component
holding a const ref to the individual fields it cares about.

Plus, I could finally get rid of debug.h/cc, which has outlived
it's usefulness.