second-pass at modularizing GUI - done?
I'm pretty happy with how SDL_GUI cleaned up. It's now pretty thin,
with just a bit of nontrivial, ANESE related behavior, namely:
1) Loading a rom passed as an arg
2) The main run loop is not fully generic (queries modules)
Neither of those are that bad though, so i'm calling it for now.

I think that after this, i'll just have 2 more tasks left until I
tag this as ANESE release 1.0.

1) Comment the codebase some more.
It's pretty good as it stands, but I want to give it a once-over.

2) See if I can't just figure out how to use SDL_QueueAudio
Yes, Blargg's Sound_Queue works _fine_, but I'd feel better if I
could say that _all_ of ANESE's sound code is hand-written.