initial tests of wideNES
I had a thought: wouldn't it be cool to "widescreenify" NES games?

There are a bunch of different ways to approach this, but i'm
experimenting with a fairly general approach: watching the PPU's
scroll registers, and shifting the screen around accordingly...

This is just an initial test, but SMB is already showing some cool
results!

As far as I can tell, i'm the first to experiment with this idea,
which is exciting, because there isn't terribly much new ground to
break in the world of NES emulators!