add padding + update READMEs
I still haven't figured out the smearing issue, but I _did_ figure
out how to greatly-improve compatibility with many titles!

I added the ability to "pad" the screen, cutting-off any artifacts
at the edge. This means that with a bit of tweaking, games like
SMB3 and MC Kids work near flawlessly!

It _should_ be possible to guess "good" padding values for games by
using some heuristic that relies on current mirroring mode, scroll
direction, and if the PPUMASK is on...

I also updated the README some more, putting a "teaser" for wideNES
on the main page, but splitting off the details into it's own page.