automatic padding detection + sampling overhaul
If a human can figure out which parts of a screen to chop, why
can't wideNES do it automatically? Well, it can! sort of...

I implemented a couple of heuristics that use the state of the PPU
and the Mapper to "guess" what padding values make sense.

For now, I only have 2 implemented:

1) Check if the bgr-mask column is enbled on PPUMASK
  - if so, add 8px of padding to the left of the screen
  - many games that mask the left also have artifacting on the
  right, so mask off 8px on the right too.

2) Check what scanline the Mapper IRQs happened on
  - mask-off the rest of the screen
  - many games use the Mapper IRQ for status-bars (smb3, kirby)

Now, these are not perfect heuristics by any means, with exceptions
rampant. eg: Lode Runner doesn't have any graphical artifacting on
the right of the screen. eg2: MegaMan IV uses the IRQ for different
reasons, unrelated to status-bars.

Nevertheless, they do a decent job with many games, smb3 and MC Kids
work flawlessly. Moreover, Metroid and Super Mario Bros work great
too, automatically having the entire screen captured (no padding)

While implemeting this feature, I decided to cleanup how I was
reading the NES's state, and implemented a basic Callback system,
enabling arbitrary methods to hook into important parts of the NES
lifecycle. eg: I added callbacks whenever a mapper fires an IRQ,
and a mapper whenever a PPU frame finished drawing (scanline 240)

This had the happy side-effect of fixing my smearing issues, as it
makes a lot more sense to sample the NES's state _before_ the NMI
handler happens, as that state the one that resulted in the current
frame! Just like that, every game I tested worked flawlessly, no
smearing to be found!

Also, this lays a much more solid groundwork for implementing a
proper debugger! It should be easy to take the nastyness that is
ppu_debug.cc and extract it into it's own GUI module! That way,
there will be _no_ state of SDL in ANESE core!